Split the broad-phase code into multiple files.
