+++ src/geometry/broad_phase_multi_sap/mod.rs
@@ -1,0 +1,16 @@
+pub use self::broad_phase::BroadPhase;
+pub use self::broad_phase_pair_event::{BroadPhasePairEvent, ColliderPair};
+
+pub(self) use self::broad_phase_proxy::*;
+pub(self) use self::sap_axis::*;
+pub(self) use self::sap_endpoint::*;
+pub(self) use self::sap_region::*;
+pub(self) use self::sap_utils::*;
+
+mod broad_phase;
+mod broad_phase_pair_event;
+mod broad_phase_proxy;
+mod sap_axis;
+mod sap_endpoint;
+mod sap_region;
+mod sap_utils;
